Completé la funcionalidad de actualizar productos. Completé la sección Productos.

+++ products/static/products/js/products.js
@@ -16,23 +16,35 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   // Agregar un event listener al checkbox "Select All"
-  selectAllCheckbox.addEventListener('change', () => {
-    rowCheckboxes.forEach((checkbox) => {
-      checkbox.checked = selectAllCheckbox.checked
+  if (selectAllCheckbox) {
+    selectAllCheckbox.addEventListener('change', () => {
+      rowCheckboxes.forEach((checkbox) => {
+        checkbox.checked = selectAllCheckbox.checked
+      });
     });
-  });
+  }
 
   // Agregar un event listener al campo de búsqueda
   const searchInput = document.querySelector('#searchInput');
   searchInput.addEventListener('input', () => {
     const filterValue = searchInput.value.toLowerCase();
     const rows = table.querySelectorAll('tbody tr');
-
+  
     rows.forEach((row) => {
-      const text = row.textContent.toLowerCase();
-      row.style.display = text.includes(filterValue) ? '' : 'none';
+      const searchableCells = row.querySelectorAll('.searchable');
+      let matchFound = false;
+  
+      searchableCells.forEach((cell) => {
+        const text = cell.textContent.toLowerCase();
+        if (text.includes(filterValue)) {
+          matchFound = true;
+        }
+      });
+  
+      row.style.display = matchFound ? '' : 'none';
     });
   });
+  
   
   // Agregar los porcentages de costo de logistica
   const tds = document.querySelectorAll('.logistics_cost_percentage');
@@ -67,35 +79,39 @@ document.addEventListener('DOMContentLoaded', () => {
     });
   });
 
-  // Elimina todos los productos seleccionados
-  deleteAllProductsButton.addEventListener("click", () => {
-    const productsCheckBoxes = document.querySelectorAll('.product-checkbox');
-    const toDelete = [];
-    productsCheckBoxes.forEach(checkBox => {
-      if (checkBox.checked) {
-        toDelete.push(checkBox);
-      }
-    });
-    if (toDelete.length > 0) {
-      swal.fire({
-        title: '¿Estás seguro?',
-        text: 'No podrás revertir esto',
-        icon: 'warning',
-        showCancelButton: true,
-        confirmButtonColor: '#3085d6',
-        cancelButtonColor: '#d33',
-        confirmButtonText: 'Sí, eliminarlo(s)'
-      }).then((result) => {
-        if (result.isConfirmed) {
-          ids = [];
-          toDelete.forEach(checkBox => {
-            ids.push(checkBox.dataset.productId);
-          });
-          deleteSelectedProducts(ids);
+  if (deleteAllProductsButton) {
+
+    // Elimina todos los productos seleccionados
+    deleteAllProductsButton.addEventListener("click", () => {
+      const productsCheckBoxes = document.querySelectorAll('.product-checkbox');
+      const toDelete = [];
+      productsCheckBoxes.forEach(checkBox => {
+        if (checkBox.checked) {
+          toDelete.push(checkBox);
         }
       });
-    }
-  })
+      if (toDelete.length > 0) {
+        swal.fire({
+          title: '¿Estás seguro?',
+          text: 'No podrás revertir esto',
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Sí, eliminarlo(s)'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            ids = [];
+            toDelete.forEach(checkBox => {
+              ids.push(checkBox.dataset.productId);
+            });
+            deleteSelectedProducts(ids);
+          }
+        });
+      }
+    });
+
+  }
 
   // Marcar el enlace Products como activo
   const $navLinks = document.querySelectorAll('.nav-link');
